pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
authors = [{ name = "TheLionCoder", email = "eorlando.reyesruiz@gmail.com" }]
name = "gdrive-suite"
version = "0.2.0"
description = "A Python library for interacting with Google Drive and Google Sheets."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Internet",
]
dependencies = [
  "google-api-python-client>=2.162.0",
  "google-auth-oauthlib>=1.2.1",
  "loguru>=0.7.3",
]

[project.urls]
"Homepage" = "https://github.com/TheLionCoder/gdrive-suite"
"Bug Tracker" = "https://github.com/TheLionCoder/gdrive-suite/issues"


[project.optional-dependencies]
dev = [
  "types-pyyaml>=6.0.12.20250516",
  "google-api-python-client-stubs",
  "google-auth-stubs",
  "build",
  "twine",
  "mkdocs>=1.6.1",
  "mkdocs-material>=9.6.18",
  "mkdocstrings-python>=1.18.2",
  "pyright>=1.1.405",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pyright]
include = ["src"]
exclude = [
  "**/node_modules",
  "**/__pycache__",
  "src/experimental",
  "src/typestubs",
]
stuPath = "src/stubs"
reportMissingImports = "error"
reportMissingStubs = false
